+++ genomepy/plugins/sizes.py
@@ -1,13 +1,17 @@
+import os.path
+
 from genomepy.plugin import Plugin
 
 class SizesPlugin(Plugin):
 
-    def after_genome_download(self, genome):
+    def after_genome_download(self, genome, force):
         props = self.get_properties(genome)
         fname = props["sizes"]
-        with open(fname, "w") as f:
-            for seqname in genome.keys():
-                f.write("{}\t{}\n".format(seqname, len(genome[seqname])))
+
+        if not os.path.exists(fname) or force is True:
+            with open(fname, "w") as f:
+                for seqname in genome.keys():
+                    f.write("{}\t{}\n".format(seqname, len(genome[seqname])))
 
     def get_properties(self, genome):
         props = {
